weather_app/requirements.txt
asgiref==3.8.1
attrs==23.2.0
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
djangorestframework==3.15.2
exceptiongroup==1.2.2
flatbuffers==24.3.25
idna==3.7
numpy==2.0.0
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.11
pandas==2.2.2
platformdirs==4.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
six==1.16.0
sqlparse==0.5.1
transliterate==1.10.2
typing_extensions==4.12.2
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.2
